% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbackdrop.R
\name{ggbackdrop}
\alias{ggbackdrop}
\title{Create a blank canvas}
\usage{
ggbackdrop(
  x = 0,
  y = 0,
  width = 100,
  alpha = 1,
  png = system.file("img", "Rlogo.png", package = "png"),
  image = png::readPNG(png),
  height = width * dim(image)[1]/dim(image)[2]
)
}
\arguments{
\item{y}{}
}
\value{

}
\description{
Create a blank canvas centered at 0, 0 to use stamps on
}
\examples{
library(ggplot2)
ggbackdrop() +
stamp_curve(curvature = .3, yend = 75, xend = 100) +
stamp_polygon(x0 = 20, y0 = 60, radius = 20, alpha = .4) +
stamp_text(x = 20, y = 60, label = "{ggstamp}",
           size = 5, color = "grey97") +
theme_void()
}
