% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcanvas.R
\name{ggcanvas}
\alias{ggcanvas}
\title{Create a blank canvas}
\usage{
ggcanvas(x = 0, y = 0)
}
\arguments{
\item{y}{}
}
\value{

}
\description{
Create a blank canvas centered at 0, 0 to use stamps on
}
\examples{
library(ggplot2)
ggcanvas() +
stamp_text() +
stamp_text(x = 2) +
theme_void() +
stamp_point(x = 1, y = 1)
}
