% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdraft.R
\name{ggdraft}
\alias{ggdraft}
\title{Create a blank canvas}
\usage{
ggdraft(x = 0, y = 0, width = 1, height = 1/1.618)
}
\arguments{
\item{y}{}
}
\value{

}
\description{
Create a blank canvas centered at 0, 0 to use stamps on
}
\examples{
library(ggplot2)
ggdraft() +
stamp_curve(curvature = .325, yend = .6) +
stamp_polygon(x0 = .2, y0 = .4, radius = .2, alpha = .4) +
stamp_text(x = .2, y = .4, label = "{ggstamp}", size = 5, color = "grey97")

last_plot() +
theme_void()
}
