% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{honeycomb}
\alias{honeycomb}
\title{Title}
\usage{
honeycomb(
  x0 = 0,
  y0 = 0,
  n = 5,
  ncol = 3,
  width = 2 * sqrt(0.75),
  height = width * 0.75/sqrt(0.75)
)
}
\arguments{
\item{height}{}
}
\value{

}
\description{
Title
}
\examples{
ggcanvas() +
 stamp_point(xy = honeycomb()) +
 stamp_polygon(x0y0 = honeycomb(n = 15, ncol = 4),
     alpha = .5) +
 stamp_label(label = 1:15, xy = honeycomb(n = 15, ncol = 4)) +
 stamp_polygon(n = 4,
               x0y0 = honeycomb(n = 15, ncol = 4)[15,],
               fill = "magenta", alpha = .4, radius = sqrt(.75))
honeycomb()
## get coords for single position...
honeycomb(n = 15, ncol = 5)[15,]

ggcanvas() +
  stamp_polygon(x0y0 = honeycomb(width = 1.8),
                fill = c("goldenrod3"), alpha = .8, size = 0)  +
  stamp_polygon(x0y0 = honeycomb(width = 1.8, n = 12)[c(6:9, 11),],
                fill = c("goldenrod4"), alpha = .8, size = 0) +
  stamp_polygon(x0y0 = honeycomb(width = 1.8, n = 15)[c(10,12:15),],
                fill = c("goldenrod2"), alpha = .8, size = 0) +
  stamp_png(x0y0 = honeycomb(width = 1.8, n = 15)[c(1,5, 10),])

}
