% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeycomb_x.R
\name{honeycomb_x}
\alias{honeycomb_x}
\title{Title}
\usage{
honeycomb_x(x0 = 0, n = 5, ncol = 3, width = 2 * sqrt(0.75))
}
\arguments{
\item{width}{}
}
\value{

}
\description{
Title
}
\examples{
honeycomb_x()

honeycomb_x(n = 10, x0 = 5, ncol = 2)
}
