% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeycomb_y.R
\name{honeycomb_y}
\alias{honeycomb_y}
\title{Title}
\usage{
honeycomb_y(y0 = 0, n = 5, ncol = 3, height = 1.5)
}
\arguments{
\item{height}{}
}
\value{

}
\description{
Title
}
\examples{
honeycomb_y()
ggcanvas() +
 stamp_polygon(x0 = honeycomb_x(),
               y0 = honeycomb_y())
}
