% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_branching.R
\name{pos_branching}
\alias{pos_branching}
\title{Title}
\usage{
pos_branching(
  x0 = 0,
  y0 = 0,
  n = 5,
  ncol = 3,
  width = 2 * sqrt(0.75),
  height = width * 0.75/sqrt(0.75)
)
}
\arguments{
\item{x0}{numeric x start position}

\item{y0}{numeric y start position}

\item{n}{number of positions}

\item{width}{numeric, width between positions}

\item{height}{numeric, height between positions}
}
\value{

}
\description{
Title
}
\examples{
# first example
pos_branching()
ggcanvas() +
 stamp_point(xy = pos_branching()) +
 stamp_polygon(x0y0 = pos_branching(n = 15, ncol = 4),
     alpha = .5) +
 stamp_label(label = 1:15, xy = pos_branching(n = 15, ncol = 4)) +
 stamp_polygon(n = 4,
               x0y0 = pos_branching(n = 15, ncol = 4)[15,],
               fill = "magenta", alpha = .4, radius = sqrt(.75))

# get coords for single position
pos_branching(n = 15)[15,]
pos_branching(n = 15)[1:4,]
}
