% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_branching_x.R
\name{pos_branching_x}
\alias{pos_branching_x}
\title{Create vector for x that, with y will, form a wrapping branching pattern}
\usage{
pos_branching_x(x0 = 0, n = 6, width = 2 * sqrt(0.75))
}
\arguments{
\item{x0}{numeric x start position}

\item{n}{number of positions}

\item{width}{numeric, width between positions}
}
\value{

}
\description{
Create vector for x that, with y will, form a wrapping branching pattern
}
\examples{
pos_branching_x()

pos_branching_x(n = 10, x0 = 5)
}
