% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_branching_y.R
\name{pos_branching_y}
\alias{pos_branching_y}
\title{Create vector for x that, with y will, form a wrapping branching pattern}
\usage{
pos_branching_y(y0 = 0, n = 12, height = 2 * sqrt(0.75))
}
\arguments{
\item{y0}{numeric y start position}

\item{n}{number of positions}

\item{height}{numeric, height between positions}
}
\value{

}
\description{
Create vector for x that, with y will, form a wrapping branching pattern
}
\examples{
pos_branching_y()

pos_branching_y(n = 10, y0 = 5)
}
