% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_days_in_month.R
\name{pos_days_in_month}
\alias{pos_days_in_month}
\title{Return positions for equilateral polygon vertices}
\usage{
pos_days_in_month(
  date = c("2020-02-11", "2020-02-15"),
  month = NULL,
  days_in_month = NULL,
  x0 = 0,
  y0 = 0,
  week_start_day = "Sunday",
  width = 1,
  height = width
)
}
\arguments{
\item{x0}{numeric for x starting point for positioner}

\item{y0}{numeric for y starting point for positioner}
}
\value{

}
\description{
Return positions for equilateral polygon vertices
}
\examples{
pos_days_in_month()
pos_days_in_month(date = "1999-02-05")
pos_days_in_month(month = "2012-01",
                  days_in_month = 2)

# Using w plotspace
ggcanvas() +
stamp_tile(xy = pos_month(month = "2020-02")) +
stamp_label(xy =
  pos_days_in_month(month = "2020-02", days_in_month = c(1,2,8)),
label = c(1,2,8))
}
