% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_honeycomb.R
\name{pos_honeycomb}
\alias{pos_honeycomb}
\title{Title}
\usage{
pos_honeycomb(
  x0 = 0,
  y0 = 0,
  n = 5,
  ncol = 3,
  width = 2 * sqrt(0.75),
  height = width * 0.75/sqrt(0.75)
)
}
\arguments{
\item{x0}{numeric x start position}

\item{y0}{numeric y start position}

\item{n}{number of positions}

\item{ncol}{number of columns before wrap, default is 3}

\item{width}{numeric, width between positions}

\item{height}{numeric, height between positions}
}
\value{

}
\description{
Title
}
\examples{
# first example
pos_honeycomb()
ggcanvas() +
 stamp_point(xy = pos_honeycomb()) +
 stamp_polygon(x0y0 = pos_honeycomb(n = 15, ncol = 4),
     alpha = .5) +
 stamp_label(label = 1:15, xy = pos_honeycomb(n = 15, ncol = 4)) +
 stamp_polygon(n = 4,
               x0y0 = pos_honeycomb(n = 15, ncol = 4)[15,],
               fill = "magenta", alpha = .4, radius = sqrt(.75))

# get coords for single position
pos_honeycomb(n = 15, ncol = 5)[15,]
pos_honeycomb(n = 15, ncol = 5)[1:4,]

# a hive
ggcanvas() +
  stamp_polygon(x0y0 = pos_honeycomb(width = 1.8),
                fill = c("goldenrod3"),
                alpha = .8,
                size = 0)  +
  stamp_polygon(x0y0 = pos_honeycomb(width = 1.8, n = 12)[c(6:9, 11),],
                fill = c("goldenrod4"),
                alpha = .8,
                size = 0) +
  stamp_polygon(x0y0 = pos_honeycomb(width = 1.8, n = 15)[c(10,12:15),],
                fill = c("goldenrod2"),
                alpha = .8,
                size = 0) +
  stamp_png(x0y0 = pos_honeycomb(width = 1.8, n = 15)[c(1,5, 10),])
}
