% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_honeycomb_x.R
\name{pos_honeycomb_x}
\alias{pos_honeycomb_x}
\title{Title}
\usage{
pos_honeycomb_x(x0 = 0, n = 5, ncol = 3, width = 2 * sqrt(0.75))
}
\arguments{
\item{width}{}
}
\value{

}
\description{
Title
}
\examples{
pos_honeycomb_x()

pos_honeycomb_x(n = 10, x0 = 5, ncol = 2)
}
