% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_honeycomb_y.R
\name{pos_honeycomb_y}
\alias{pos_honeycomb_y}
\title{Create vector for y that, with x will, form a wrapping honeycomb pattern}
\usage{
pos_honeycomb_y(y0 = 0, n = 5, ncol = 3, height = 1.5)
}
\arguments{
\item{y0}{numeric y start position}

\item{n}{number of positions}

\item{ncol}{number of columns before wrap, default is 3}

\item{height}{numeric, height between positions}
}
\value{

}
\description{
Create vector for y that, with x will, form a wrapping honeycomb pattern
}
\examples{
pos_honeycomb_y()
ggcanvas() +
 stamp_polygon(x0 = pos_honeycomb_x(),
               y0 = pos_honeycomb_y())
}
