% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_month.R
\name{pos_month}
\alias{pos_month}
\title{Return positions for equilateral polygon vertices}
\usage{
pos_month(
  month = "2021-10",
  x0 = 0,
  y0 = 0,
  week_start_day = "Sunday",
  width = 1,
  height = width
)
}
\arguments{
\item{x0}{numeric for x starting point for positioner}

\item{y0}{numeric for y starting point for positioner}
}
\value{

}
\description{
Return positions for equilateral polygon vertices
}
\examples{
pos_month()
pos_month(month = "1999-02")

# Using w plotspace
ggcanvas() +
stamp_tile(xy = pos_month("2021-10")) +
stamp_text(xy = pos_month("2021-10"),
label = 1:31) +
stamp_point(alpha = .2,
xy = pos_month()[3:12,],
color = "blue", size = 11) +
stamp_tile(xy =
pos_month("2000-02",
x0 = 0, y0 = -5),
fill = "linen")
}
