% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_polygon.R
\name{pos_polygon}
\alias{pos_polygon}
\title{Return positions for equilateral polygon vertices}
\usage{
pos_polygon(x0 = 0, y0 = 0, n = 5, radius = 1, rotation = -0.5)
}
\arguments{
\item{x0}{numeric x center of shape, defaults to 0}

\item{y0}{numeric y center of shape, defaults to 0}

\item{n}{number of positions}

\item{radius}{a numeric value, the distance between x0 and y0 and the vertices}

\item{rotation}{numeric indicating how much polygon should be turned from first vertex along positive horizontal}
}
\value{

}
\description{
Return positions for equilateral polygon vertices
}
\examples{
pos_polygon()
pos_polygon(n = 10)

# Using w plotspace
ggcanvas() +
stamp_label(label = c("hi", "bye", "tchao"),
            xy = pos_polygon(n = 3,
            rotation = -.5,
            radius = .5))
}
