% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_polygon_x.R
\name{pos_polygon_x}
\alias{pos_polygon_x}
\title{Title}
\usage{
pos_polygon_x(x0 = 0, n = 5, radius = 1, rotation = -0.5)
}
\arguments{
\item{x0}{numeric x center of shape, defaults to 0}

\item{radius}{a numeric value, the distance between x0 and y0 and the vertices}

\item{rotation}{numeric indicating how much polygon should be turned from first vertex along positive horizontal}
}
\value{

}
\description{
help with documentation
}
\examples{
pos_polygon_x()
pos_polygon_x(n = 10)
}
