% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_polygon_y.R
\name{pos_polygon_y}
\alias{pos_polygon_y}
\title{Create vector for y that, with x will, form a polygon pattern}
\usage{
pos_polygon_y(y0 = 0, n = 5, radius = 1, rotation = -0.5)
}
\arguments{
\item{y0}{numeric y center of shape, defaults to 0}

\item{radius}{a numeric value, the distance between x0 and y0 and the vertices}

\item{rotation}{numeric indicating how much polygon should be turned from first vertex along positive horizontal}
}
\value{

}
\description{
Create vector for y that, with x will, form a polygon pattern
}
\examples{
pos_polygon_y(n = 6)
ggcanvas() +
stamp_label(label = "hello",
           x = pos_polygon_x(n = 5),
           y = pos_polygon_y(n = 5)) +
stamp_point(x = pos_polygon_x(n = 5, rot = 0),
            y = pos_polygon_y(n = 5, rot = 0))
}
