% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_spoke.R
\name{pos_spoke}
\alias{pos_spoke}
\title{Title}
\usage{
pos_spoke(x0 = 0, y0 = 0, n = 5, radius = 1, rotation = -0.5)
}
\arguments{
\item{rotation}{}
}
\value{

}
\description{
Title
}
\examples{
pos_spoke()
pos_spoke(n = 10)

# Using w plotspace
ggcanvas() +
stamp_label(label = c("hi", "bye", "tchao"),
            xy = pos_spoke(n = 3,
            rotation = -.5,
            radius = .5))
}
