% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_spoke_x.R
\name{pos_spoke_x}
\alias{pos_spoke_x}
\title{Title}
\usage{
pos_spoke_x(x0 = 0, n = 5, radius = 1, rotation = -0.5)
}
\arguments{
\item{rotation}{}
}
\value{

}
\description{
help with documentation
}
\examples{
pos_spoke_x()
pos_spoke_x(n = 10)
}
