% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_spoke_y.R
\name{pos_spoke_y}
\alias{pos_spoke_y}
\title{Title}
\usage{
pos_spoke_y(y0 = 0, n = 5, radius = 1, rotation = -0.5)
}
\arguments{
\item{y0}{}
}
\value{

}
\description{
Title
}
\examples{
pos_spoke_y(n = 6)
ggcanvas() +
stamp_label(label = "hello",
           x = pos_spoke_x(n = 5),
           y = pos_spoke_y(n = 5)) +
stamp_point(x = pos_spoke_x(n = 5, rot = 0),
            y = pos_spoke_y(n = 5, rot = 0))
}
