% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_stack.R
\name{pos_stack}
\alias{pos_stack}
\title{Title}
\usage{
pos_stack(x0 = 0, y0 = 0, n = 5, ncol = 3, width = 1, height = width)
}
\arguments{
\item{ncol}{}
}
\value{

}
\description{
Title
}
\examples{
ggcanvas() +
 stamp_point(xy = pos_stack(), color = "blue") +
 stamp_text(xy = pos_stack(), label = letters[1:5])

}
