% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_stack_x.R
\name{pos_stack_x}
\alias{pos_stack_x}
\title{Title}
\usage{
pos_stack_x(x0 = 0, n = 5, ncol = 3, width = 1)
}
\arguments{
\item{x0}{numeric for x starting point for positioner}

\item{n}{number of positions}

\item{ncol}{number of columns before wrap}
}
\value{

}
\description{
Title
}
