% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_wrap.R
\name{pos_wrap}
\alias{pos_wrap}
\title{Title}
\usage{
pos_wrap(x0 = 0, y0 = 0, n = 5, ncol = 3, width = 1, height = width)
}
\arguments{
\item{ncol}{}
}
\value{

}
\description{
Title
}
\examples{
pos_wrap()
pos_wrap(n = 7, ncol = 2)
}
