% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_wrap_multi.R
\name{pos_wrap_multi}
\alias{pos_wrap_multi}
\title{Title}
\usage{
pos_wrap_multi(
  x0 = 0,
  y0 = 0,
  n = c(5, 8),
  ncol = 3,
  width = 1,
  spacing = 1,
  width_multi = ncol * width + spacing,
  height = width
)
}
\arguments{
\item{x0}{numeric for x starting point for positioner}

\item{y0}{numeric for y starting point for positioner}

\item{n}{number of positions}

\item{ncol}{number of columns before wrap}

\item{spacing}{width between wrapped groups, defaulst to 1}

\item{width_multi}{spacing for groups, defaults to width + spacing}
}
\value{

}
\description{
Title
}
\examples{
# just printing
pos_wrap_multi(n = c(3, 4), ncol = 2)
pos_wrap_multi()

# in plot
ggcanvas() +
  stamp_point(xy =
      pos_wrap_multi(n = c(5, 10),
                     ncol = 4))
}
