% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_wrap_snake.R
\name{pos_wrap_snake}
\alias{pos_wrap_snake}
\title{Title}
\usage{
pos_wrap_snake(x0 = 0, y0 = 0, n = 5, ncol = 3, width = 1, height = width)
}
\arguments{
\item{x0}{numeric for x starting point for positioner}

\item{y0}{numeric for y starting point for positioner}

\item{n}{number of positions}

\item{ncol}{number of columns before wrap}
}
\value{

}
\description{
Title
}
\examples{
pos_wrap_snake()

ggcanvas() +
 stamp_path(xy = pos_wrap_snake(n = 9, ncol = 3)) +
 stamp_label(xy = pos_wrap_snake(n = 9, ncol = 3),
 label = LETTERS[1:9])
}
