% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_wrap_x.R
\name{pos_wrap_x}
\alias{pos_wrap_x}
\title{Title}
\usage{
pos_wrap_x(x0 = 0, n = 5, ncol = 3, width = 1, snake = FALSE)
}
\arguments{
\item{x0}{numeric for x starting point for positioner}

\item{n}{number of positions}

\item{ncol}{number of columns before wrap}
}
\value{

}
\description{
Title
}
\examples{
pos_wrap_x()
pos_wrap_x(n = 12)
pos_wrap_x(snake = TRUE)
}
