% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_wrap_png_x.R
\name{pos_wrap_x_png}
\alias{pos_wrap_x_png}
\title{Title}
\usage{
pos_wrap_x_png(
  x0 = 0,
  n = 5,
  ncol = 3,
  png = system.file("img", "Rlogo.png", package = "png"),
  image = png::readPNG(png),
  width = 1
)
}
\arguments{
\item{x0}{numeric for x starting point for positioner}

\item{n}{number of positions}

\item{ncol}{number of columns before wrap}

\item{png}{a string with the path to an image, defaults to R image}

\item{image}{a numeric array with png red green blue alpha information, defaults to png::readPNG(png)}

\item{width}{the width that the image should fill, defaults to 1}
}
\value{

}
\description{
Title
}
