% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_wrap_png_y.R
\name{pos_wrap_y_png}
\alias{pos_wrap_y_png}
\title{Title}
\usage{
pos_wrap_y_png(
  y0 = 0,
  n = 5,
  ncol = 3,
  png = system.file("img", "Rlogo.png", package = "png"),
  image = png::readPNG(png),
  width = 1,
  height = width * dim(image)[2]/dim(image)[1]
)
}
\arguments{
\item{y0}{numeric for y starting point for positioner}

\item{n}{number of positions}

\item{ncol}{number of columns before wrap}

\item{png}{a string with the path to an image, defaults to R image}

\item{image}{a numeric array with png red green blue alpha information, defaults to png::readPNG(png)}

\item{width}{the width that the image should fill, defaults to 1}

\item{height}{the height that the image should fill, calculated to match aspect ratio based on width}
}
\value{

}
\description{
Title
}
