% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue_split.R
\name{pvalue_split}
\alias{pvalue_split}
\title{Format p values to charactor}
\usage{
pvalue_split(data)
}
\arguments{
\item{data}{a p value vector}
}
\value{
a data frame
}
\description{
Format p values to charactor
}
\examples{
pvalue_split(c("2.27037499009e-22", "7.18282161205e-03"))
}
\author{
Zhonghui Gai
}
