% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{spoke_y}
\alias{spoke_y}
\title{Title}
\usage{
spoke_y(y0 = 0, n = 5, radius = 1, rotation = -0.5)
}
\arguments{
\item{y0}{}
}
\value{

}
\description{
Title
}
\examples{
spoke_y(n = 6)
ggcanvas() +
stamp_label(label = "hello",
           x = spoke_x(n = 5),
           y = spoke_y(n = 5)) +
stamp_point(x = spoke_x(n = 5, rot = 0),
            y = spoke_y(n = 5, rot = 0)) +
stamp_label(label = c("hi", "bye", "chao"),
            x = spoke_x(n = 3, rotation = -.5, radius = .5),
            y = spoke_y(n = 3, rotation = -.5, radius = .5))
}
