% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_bar.R
\name{stamp_bar}
\alias{stamp_bar}
\title{Draw the bar plot}
\usage{
stamp_bar(data, legend.position = c(-0.8, 1.025), group = NULL)
}
\arguments{
\item{data}{the list from stamp_input function}

\item{legend.position}{legend.position}

\item{group}{color information}
}
\value{
a ggplot2 object
}
\description{
Draw the bar plot
}
\examples{
group <- c("#E69F00", "#56B4E9")
stamp_bar(stamp_input(data = stamp_input(), group = group))
}
\author{
Zhonghui Gai
}
