% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_bubble.R
\name{stamp_bubble}
\alias{stamp_bubble}
\title{Stamp some text}
\usage{
stamp_bubble(
  x = 0,
  y = 0,
  size = 8,
  stroke = 2,
  shape = 21,
  alpha = 1,
  fill = "grey85",
  color = "black"
)
}
\arguments{
\item{x}{defaults to 0}

\item{y}{defaults to 0}

\item{size}{}

\item{label}{defaults with a message about used}
}
\value{

}
\description{
This function adds a text annotation layer
}
\examples{
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 stamp_bubble() +
 stamp_bubble(color = "red", x = 10, y = 100)

ggcanvas() +
 stamp_bubble(color = "green") +
 stamp_bubble(x = 1:5, y = 2, size = 8)
}
