% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_CI.R
\name{stamp_CI}
\alias{stamp_CI}
\title{Draw the CI plot of STAMP}
\usage{
stamp_CI(data, group = NULL)
}
\arguments{
\item{data}{the output of stamp_input()}

\item{group}{color information}
}
\value{
a ggplot2 object
}
\description{
Draw the CI plot of STAMP
}
\examples{
stamp_CI(stamp_input())
}
\author{
Zhonghui Gai
}
