% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_circle.R
\name{stamp_circle}
\alias{stamp_circle}
\title{Stamp a polygon of 100 sides}
\usage{
stamp_circle(
  x0 = 0,
  y0 = 0,
  n = 100,
  radius = 1,
  size = 1.5,
  alpha = 1,
  rotation = -0.5,
  linetype = "solid",
  fill = "grey35",
  color = "black",
  x0y0 = NULL
)
}
\arguments{
\item{x0}{defaults to 0}

\item{y0}{defaults to 0}

\item{color}{}
}
\value{

}
\description{
This function adds a text annotation layer.
Contribute or help document https://github.com/EvaMaeRey/ggstamp/blob/master/R/stamp_polygon.R
}
\examples{
# on plot space
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_circle(color = "darkred", radius = 12)

# on a canvas
 ggcanvas() +
  stamp_circle(x0 = c(0, 2.5),
  fill = "magenta", color = "purple") +
  stamp_label(xy = pos_spoke(x0 = 0, n = 3),
              label = c("hi", "hello", "bye")) +
  stamp_point(color = "turquoise")

# lots of circles
 ggcanvas() +
 stamp_circle(color = "turquoise4", radius = 3, fill = "turquoise") +
 stamp_circle(fill = "slateblue2", radius = 2) +
 stamp_circle(linetype = "dashed", alpha = .5, fill = "plum2") +
 stamp_circle(x0 = 1, y0 = 1.5, rotation = .25, size = 0,
  fill = "goldenrod2") +
 stamp_circle(n = 50, radius = .2, size = 0) +
 stamp_spoke(angle = pi * 1:7/15, radius = 2)

# some more
 ggcanvas() +
  stamp_circle(x0y0 = pos_wrap(n = 5, ncol = 3, height = 1.5),
               radius = .55)

}
