% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_hline.R
\name{stamp_hline}
\alias{stamp_hline}
\title{Stamp a horizontal line, a wrapper for geom_hline}
\usage{
stamp_hline(
  yintercept = 0,
  alpha = 1,
  color = "black",
  linetype = "solid",
  size = 1.5
)
}
\arguments{
\item{yintercept}{numeric indicating yintercept, default to 0}

\item{alpha}{numeric between 0 and 1 indicating transparency, defaults to 1 fully opaque, zero is fully transparent}

\item{color}{a character string indicating color, defaults to "black", for other named colors check out https://evamaerey.github.io/ggplot2_grammar_guide/named_colors}

\item{linetype}{a string indicating linetype, "solid", "dashed", "dotted", defaults to "solid"}

\item{size}{numeric value indicating line width, defaults to 1.5}
}
\value{

}
\description{
Stamp a horizontal line, a wrapper for geom_hline
}
\examples{
library(ggplot2)
ggplot(cars) +
aes(speed, dist) +
stamp_hline()

# on a canvas
ggcanvas() +
stamp_hline() +
stamp_hline(1:6,
           linetype = rep(c("dotted", "dashed"), 3))
}
