% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_image.R
\name{stamp_image}
\alias{stamp_image}
\title{stamp_image}
\usage{
stamp_image(
  image = png::readPNG(system.file("img", "Rlogo.png", package = "png")),
  grob = grid::rasterGrob(image = image),
  x = 0,
  y = 0,
  width = 1,
  height = width * dim(image)[2]/dim(image)[1],
  show.legend = F
)
}
\arguments{
\item{ymax}{}
}
\value{

}
\description{
stamp_image
}
\examples{
ggplot(cars, aes(dist, speed)) +
  geom_point() +
  stamp_image(width = 10, x = 40, y = 20, height = 1)

ggcanvas() +
  stamp_image() +
  stamp_image(x = 2:3, y = 6, width = .95) +
  coord_equal() +
  scale_fill_viridis_c()

ggcanvas() +
  stamp_image(x = 0:8 \%\% 5 + 1,
        y = -(0:8 \%/\% 5)/.75 , width = .90) +
  coord_equal() +
  stamp_image(x = 0:5 \%\% 5 + 1,
      y = -(0:5 \%/\% 5)/.75  -3, width = .90) +
  stamp_image(x = 0:2 \%\% 5 + 1,
     y = -(0:2 \%/\% 5)/.75  -6, width = .90)
}
