% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_input.R
\name{stamp_input}
\alias{stamp_input}
\title{Readin the data.txt file, and trim it.}
\usage{
stamp_input(path = "stamp.txt")
}
\arguments{
\item{path}{the path for the data form STAMP}
}
\value{
a list containing 2 data.frame, 1 for fig1, 1 for fig2-3.
}
\description{
Readin the data.txt file, and trim it.
}
\examples{
stamp_input("stamp.txt")
}
\author{
ZHonghui Gai
}
