% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_patch.R
\name{stamp_patch}
\alias{stamp_patch}
\title{Patch figures}
\usage{
stamp_patch(p.bar, p.CI, p.values)
}
\arguments{
\item{p.bar}{the output of stamp_bar}

\item{p.CI}{the output of stamp_CI}

\item{p.values}{the output of stamp_Point}
}
\value{
a patched ggplots object
}
\description{
Patch figures
}
\examples{
stamp_patch(stamp_bar(stamp_input()), stamp_CI(stamp_input()), stamp_pvalue_point(stamp_input()))
}
