% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_path.R
\name{stamp_path}
\alias{stamp_path}
\title{Stamp some text}
\usage{
stamp_path(
  x = 0:3,
  y = c(0, 1, 0, 1),
  alpha = 1,
  color = "black",
  linetype = "solid",
  size = 1.5,
  lineend = "butt",
  linejoin = "round",
  arrow = arrow(length = 0),
  arrow.fill = NULL,
  xy = NULL
)
}
\arguments{
\item{x}{defaults to 0}

\item{y}{defaults to 0}

\item{alpha}{a numeric between 0 and 1 for transparency, zero is fully transparent, 1 fully opaque}

\item{color}{line color}

\item{linetype}{"solid", "dashed", "dotted"}

\item{size}{numeric indicating line width, defaults to 1.5}

\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{arrow}{specification for arrow heads, as created by arrow().}

\item{arrow.fill}{fill colour to use for the arrow head (if closed). NULL means use colour aesthetic.}
}
\value{

}
\description{
This function adds a text annotation layer
}
\examples{
# segments on canvas
 ggcanvas() +
 stamp_path(color = "green") +
 stamp_path(xy = pos_honeycomb(ncol = 1, n = 12),
            alpha = .5, lineend = "round")


}
