% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_polygon.R
\name{stamp_polygon}
\alias{stamp_polygon}
\title{Stamp a polygon of n sides}
\usage{
stamp_polygon(
  x = c(0, 1, 0, 1),
  y = c(0, 0, 1, 1),
  x0 = 0,
  y0 = 0,
  n = 6,
  radius = 1,
  size = 1.5,
  alpha = 1,
  rotation = 0.5,
  linetype = "solid",
  fill = "grey35",
  color = "black"
)
}
\arguments{
\item{x0}{defaults to 0}

\item{y0}{defaults to 0}

\item{size}{}

\item{label}{defaults with a message about used}
}
\value{

}
\description{
This function adds a text annotation layer
}
\examples{
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_polygon(color = "darkred")

 ggcanvas() +
 stamp_polygon(color = "turquoise4", radius = 3, n = 5, fill = "turquoise") +
 stamp_polygon(fill = "slateblue2", radius = 2, n = 7) +
 stamp_polygon(linetype = "dashed", alpha = .5, fill = "plum2") +
 stamp_polygon(x0 = 1, y0 = 1.5, n = 3, rotation = .25, size = 0,
  fill = "goldenrod2") +
 stamp_polygon(n = 50, radius = .2, size = 0) +
 stamp_spoke(angle = pi * 1:7/15, radius = 2)
}
