% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_polygon_inverse.R
\name{stamp_polygon_inverse}
\alias{stamp_polygon_inverse}
\title{Stamp a polygon of n sides}
\usage{
stamp_polygon_inverse(
  x0 = 0,
  y0 = 0,
  n = 6,
  radius = 1,
  radius_outer = radius * 2^0.5,
  size = 0,
  alpha = 1,
  rotation = -0.5,
  linetype = "solid",
  fill = "white",
  color = "black",
  x0y0 = NULL
)
}
\arguments{
\item{x0}{defaults to 0}

\item{y0}{defaults to 0}

\item{color}{}
}
\value{

}
\description{
This function adds a text annotation layer.
Contribute or help document https://github.com/EvaMaeRey/ggstamp/blob/master/R/stamp_polygon.R
}
\examples{
# on plot space
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_polygon_inverse(x0 = 20, y0 =50, color = "darkred",
 size = .5,
 radius = 10)

# on a canvas
 ggcanvas() +
  stamp_polygon_inverse(n = 3, x0 = c(0,2.5),
  fill = "magenta", color = "purple") +
  stamp_label(x = pos_spoke_x(x0 = 0, n = 3),
              y = pos_spoke_y(y0 = 0, n = 3),
              label = c("hi", "hello", "bye")) +
  stamp_point(color = "turquoise")

# lots of polygons
 ggcanvas() +
 stamp_polygon_inverse(radius = 3, n = 5,
                       fill = "turquoise") +
 stamp_polygon_inverse(fill = "slateblue2",
                       radius = 1.5, n = 7) +
 stamp_polygon_inverse(linetype = "dashed",
                       alpha = .5, fill = "plum2") +
 stamp_polygon_inverse(x0 = 1, y0 = 1.5, n = 3,
                       rotation = .25, size = 0,
                       fill = "goldenrod2") +
 stamp_polygon_inverse(n = 50, radius = .2,
                       size = 0) +
 stamp_spoke(angle = pi * 1:7/15, radius = 2)

# some more
 ggcanvas() +
  stamp_png() +
  stamp_polygon_inverse(radius = .45,
                        radius_outer = Inf) +
  stamp_polygon(alpha = .2, radius = .45,
                color = "grey30")
}
