% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_pvalue_point.R
\name{stamp_pvalue_point}
\alias{stamp_pvalue_point}
\title{Draw pvalue point}
\usage{
stamp_pvalue_point(data)
}
\arguments{
\item{data}{the output of stamp_input function}
}
\value{
a ggplot2 object
}
\description{
Draw pvalue point
}
\examples{
stamp_pvalue_point(stamp_input("stamp.txt"))
}
\author{
Zhonghui Gai
}
