% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_pvalue_text.R
\name{stamp_pvalue_text}
\alias{stamp_pvalue_text}
\title{Draw p value text}
\usage{
stamp_pvalue_text(data)
}
\arguments{
\item{data}{list from stamp_input function}
}
\value{
a ggplot2 object
}
\description{
Draw p value text
}
\examples{
stamp_pvalue_text(stamp_input())
}
\author{
Zhonghui Gai
}
