% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_rect.R
\name{stamp_rect}
\alias{stamp_rect}
\title{Stamp some text}
\usage{
stamp_rect(
  xmin = -1,
  ymin = -1,
  xmax = 1,
  ymax = 1,
  x = 0,
  y = 0,
  alpha = 1,
  color = "black",
  fill = "grey85",
  linetype = "solid",
  size = 0.5,
  xxyy = NULL
)
}
\arguments{
\item{x}{defaults to 0}

\item{y}{defaults to 0}

\item{alpha}{0 to 1 indicating transparency}

\item{color}{named or hexcolor for outline of rect}

\item{fill}{background color of label, consider named colors https://evamaerey.github.io/ggplot2_grammar_guide/named_colors}

\item{linetype}{"solid" "dashed" "dotted" linetype border}

\item{size}{line width default is .5}
}
\value{

}
\description{
This function adds a text annotation layer
}
\examples{
# rect
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_rect(color = "red") +
 stamp_rect(
            xxyy = data.frame(xmin = 1:4*3, xmax = 2:5*3,
            ymin = 1:4*6, ymax = 1:8*6))

# on a canvas
ggcanvas() +
 stamp_rect(fill = "green") +
 stamp_rect(alpha = .8,
            fill = "steelblue",
            size = 0,
            xmin = 2, xmax = 3,
            ymin = 5, ymax = 8) +
 coord_equal()
}
