% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_text_parse.R
\name{stamp_text_parse}
\alias{stamp_text_parse}
\title{Stamp some text}
\usage{
stamp_text_parse(
  x = 0,
  y = 0,
  label = "bold(italic(p)*\\"-value\\"<\\"0.01\\")",
  text_wrap = FALSE,
  char_width = 40,
  alpha = 1,
  angle = 0,
  color = "black",
  family = "Times",
  fontface = "bold",
  hjust = 0.5,
  lineheight = 0.85,
  size = 8,
  vjust = 0.5,
  xy = NULL,
  parse = TRUE,
  nudge_x = 0,
  nudge_y = 0
)
}
\arguments{
\item{label}{defaults with a message about use}

\item{alpha}{numeric between 0 and 1 indicating transparency, defaults to 1 fully opaque, zero is fully transparent}

\item{color}{a character string indicating color, defaults to "black", for other named colors check out https://evamaerey.github.io/ggplot2_grammar_guide/named_colors}

\item{family}{font family}

\item{hjust}{horizontal justification, between 0 and 1, defaults to .5}

\item{lineheight}{spacing between lines if text is wrapped}

\item{size}{size of text}

\item{vjust}{vertical justification, between 0 and 1, defaults to .5}

\item{xy}{a data frame where the first two columns contain x and y positional values}

\item{parse}{If TRUE, the labels will be parsed into expressions and displayed as described in ?plotmath.}

\item{nudge_x}{Horizontal adjustment to nudge labels by. Useful for offsetting text from points, particularly on discrete scales.}

\item{nudge_y}{Vertical adjustment to nudge labels by. Useful for offsetting text from points, particularly on discrete scales.}
}
\value{

}
\description{
This function adds a text annotation layer
}
\examples{
# stamping text on a plot
library(ggplot2)
ggcanvas() +
 stamp_text_parse()

}
