% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_text_box.R, R/stamp_textbox.R
\name{stamp_textbox}
\alias{stamp_textbox}
\title{Stamp some text}
\usage{
stamp_textbox(
  x = 0,
  y = 0,
  label = "a *very* nice **label**",
  alpha = 1,
  color = "black",
  fill = "white",
  family = "Times",
  hjust = 0.5,
  lineheight = 0.9,
  size = 8,
  vjust = 0.5,
  orientation = "upright",
  width = unit(0.8, "npc"),
  nudge_x = 0,
  nudge_y = 0,
  label.padding = ggplot2::unit(0.5, units = "lines"),
  label.r = ggplot2::unit(0.15, units = "lines"),
  label.size = 0.3,
  parse = FALSE,
  xy = NULL
)

stamp_textbox(
  x = 0,
  y = 0,
  label = "a *very* nice **label**",
  alpha = 1,
  color = "black",
  fill = "white",
  family = "Times",
  hjust = 0.5,
  lineheight = 0.9,
  size = 8,
  vjust = 0.5,
  orientation = "upright",
  width = unit(0.8, "npc"),
  nudge_x = 0,
  nudge_y = 0,
  label.padding = ggplot2::unit(0.5, units = "lines"),
  label.r = ggplot2::unit(0.15, units = "lines"),
  label.size = 0.3,
  parse = FALSE,
  xy = NULL
)
}
\arguments{
\item{x}{defaults to 0}

\item{y}{defaults to 0}

\item{label}{defaults with a message about use}

\item{alpha}{transparency of label}

\item{color}{color of text, consider named colors https://evamaerey.github.io/ggplot2_grammar_guide/named_colors}

\item{fill}{background color of label, consider named colors https://evamaerey.github.io/ggplot2_grammar_guide/named_colors}

\item{family}{font family}

\item{hjust}{horizontal justification, between 0 and 1, defaults to .5}

\item{lineheight}{spacing between lines if text is wrapped}

\item{size}{size of text}

\item{vjust}{vertical justification, between 0 and 1, defaults to .5}

\item{width}{}

\item{nudge_x}{Horizontal adjustment to nudge labels by. Useful for offsetting text from points, particularly on discrete scales.}

\item{nudge_y}{Vertical adjustment to nudge labels by. Useful for offsetting text from points, particularly on discrete scales.}

\item{label.padding}{Amount of padding around label. Defaults to 0.5 lines.}

\item{label.r}{Radius of rounded corners. Defaults to 0.15 lines.}

\item{label.size}{Size of label border, in mm.}

\item{parse}{If TRUE, the labels will be parsed into expressions and displayed as described in ?plotmath.}

\item{xy}{a data frame where the first two columns contain x and y positional values}
}
\value{



}
\description{
This function adds a text annotation layer

This function adds a text annotation layer
}
\examples{
# labels on a plot
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_label(x = 20, y = 10) +
 stamp_label(label = "hi",
 y = c(100, 75, 50), x = 5, fill = "cadetblue2", label.size = 0)

# label on a canvas
ggcanvas() +
   stamp_label()
# labels on a plot
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_textbox(x = 20, y = 40) +
 stamp_textbox(x = 20, y = 20, label = "*hello*") +
 stamp_textbox(label = "**world**", hjust = 0,
 y = 10, x = 5, fill = "cadetblue2", label.size = 0) +
 stamp_point()

# label on a canvas
library(ggtext)
ggdraft() +
   stamp_textbox(40, 30, width = unit(0.3, "npc")) +
   stamp_point(x = 1:3*10, y = 1)

}
