% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_tile.R
\name{stamp_tile}
\alias{stamp_tile}
\title{Stamp some text}
\usage{
stamp_tile(x = 0, y = 0, size = 2, alpha = 1, color = "black")
}
\arguments{
\item{x}{defaults to 0}

\item{y}{defaults to 0}

\item{size}{}

\item{label}{defaults with a message about used}
}
\value{

}
\description{
This function adds a text annotation layer
}
\examples{
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_tile(color = "red", x = 10, y = 100)

 ggcanvas() +
 stamp_tile(color = "green") +
 stamp_tile(x = 1:5, y = 2, size = 3, alpha = .5) +
 coord_equal()
}
