% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_vline.R
\name{stamp_vline}
\alias{stamp_vline}
\title{Horizontal line, wraps geom_vline w defaut at y = 0}
\usage{
stamp_vline(
  xintercept = 0,
  alpha = 1,
  color = "black",
  linetype = "solid",
  size = 1.5
)
}
\arguments{
\item{size}{}
}
\value{

}
\description{
Horizontal line, wraps geom_vline w defaut at y = 0
}
\examples{
library(ggplot2)
ggplot(cars) +
aes(speed, dist) +
geom_point() +
stamp_vline()

# on a canvas
ggcanvas() +
stamp_vline(color = "cadetblue") +
stamp_vline(0:5,
           linetype = rep(c("dotted", "dashed"), 3))
}
