% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_wash.R
\name{stamp_wash}
\alias{stamp_wash}
\title{Stamp some text}
\usage{
stamp_wash(
  xmin = -Inf,
  ymin = -Inf,
  xmax = Inf,
  ymax = Inf,
  x = 0,
  y = 0,
  alpha = 0.5,
  color = alpha("linen", 0),
  fill = "linen",
  linetype = "solid",
  size = 0.5,
  xxyy = NULL
)
}
\arguments{
\item{alpha}{numeric between 0 and 1 indicating transparency, defaults to 1 fully opaque, zero is fully transparent}

\item{color}{a character string indicating color, defaults to "black", for other named colors check out https://evamaerey.github.io/ggplot2_grammar_guide/named_colors}

\item{fill}{a character string indicating fill color, defaults to "grey35", for other named colors check out https://evamaerey.github.io/ggplot2_grammar_guide/named_colors}

\item{linetype}{a string indicating linetype, "solid", "dashed", "dotted", defaults to "solid"}

\item{size}{numeric value indicating line width, defaults to 1.5}
}
\value{

}
\description{
This function adds a text annotation layer
}
\examples{
# wash
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_wash() +
 stamp_wash(fill = "red")

# over backdrop
ggbackdrop() +
 stamp_wash(fill = "green",
            alpha = .2)
}
