% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_stamp.R
\name{theme_stamp}
\alias{theme_stamp}
\title{ggplot theme for publication ready Plots, adapted from ggtheme::theme_pub}
\usage{
theme_stamp(base_size = 13, base_family = "serif", border = FALSE, angle = 0)
}
\arguments{
\item{base_size}{the default value is 14}

\item{base_family}{the default value is "serif"}

\item{border}{`logical`. Should a border be drawn around the plot?}

\item{angle}{`integer`. Angle of x_axis text in degrees.
One of: `-30, 45, 90, 270`.}
}
\value{
ggplot theme
}
\description{
ggplot theme for publication ready Plots, adapted from ggtheme::theme_pub
}
\examples{
gglm::gglm() + ggtheme::theme_stamp()
gglm(group = "Species", label.y = 0.99) + ggtheme::theme_stamp() +
ggprism::scale_color_prism(palette = "floral")
}
\author{
Zhonghui Gai
}
