% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_void_fill.R
\name{theme_void_fill}
\alias{theme_void_fill}
\title{Title}
\usage{
theme_void_fill(fill = "linen", color = fill)
}
\arguments{
\item{color}{}
}
\value{

}
\description{
Title
}
\examples{
ggcanvas() +
  stamp_png() +
  theme_void_fill()
}
