% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{trim black before and after a variable}
\usage{
trim(x)
}
\arguments{
\item{x}{the charactor vector}
}
\value{
a trimed vector
}
\description{
trim black before and after a variable
}
\examples{
x <- c("test ", " trim ")
trim(x)
}
\author{
Zhonghui Gai
}
