% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{wrap}
\alias{wrap}
\title{Title}
\usage{
wrap(x0 = 0, y0 = 0, n = 5, ncol = 3, width = 1, height = width)
}
\arguments{
\item{ncol}{}
}
\value{

}
\description{
Title
}
\examples{
wrap()
wrap(n = 7, ncol = 2)
}
