% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_png.R
\name{wrap_png}
\alias{wrap_png}
\title{Title}
\usage{
wrap_png(
  x0 = 0,
  y0 = 0,
  n = 5,
  ncol = 3,
  png = system.file("img", "Rlogo.png", package = "png"),
  image = png::readPNG(png),
  width = 1,
  height = width * dim(image)[1]/dim(image)[2]
)
}
\arguments{
\item{height}{}
}
\value{

}
\description{
Title
}
\examples{
ggcanvas() +
    stamp_png(x0y0 = wrap_png(n = 7)) +
    stamp_png(x0y0 = wrap_png(n = 5,
              y0 = -2.5), width = .85)
}
