% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_png.R
\name{wrap_x_png}
\alias{wrap_x_png}
\title{Title}
\usage{
wrap_x_png(
  x0 = 0,
  n = 5,
  ncol = 3,
  png = system.file("img", "Rlogo.png", package = "png"),
  image = png::readPNG(png),
  width = 1
)
}
\arguments{
\item{n}{}
}
\value{

}
\description{
Title
}
