% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_png_y.R
\name{wrap_y_png}
\alias{wrap_y_png}
\title{Title}
\usage{
wrap_y_png(
  y0 = 0,
  n = 5,
  ncol = 3,
  png = system.file("img", "Rlogo.png", package = "png"),
  image = png::readPNG(png),
  width = 1,
  height = width * dim(image)[2]/dim(image)[1]
)
}
\arguments{
\item{n}{}
}
\value{

}
\description{
Title
}
