% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbackdrop.R
\name{ggbackdrop}
\alias{ggbackdrop}
\title{Create a blank canvas}
\usage{
ggbackdrop(
  x = 0,
  y = 0,
  width = 1,
  alpha = 1,
  png = system.file("img", "Rlogo.png", package = "png"),
  image = png::readPNG(png),
  height = width * dim(image)[1]/dim(image)[2]
)
}
\arguments{
\item{y}{}
}
\value{

}
\description{
Create a blank canvas centered at 0, 0 to use stamps on
}
\examples{
library(ggplot2)
ggbackdrop() +
stamp_curve(curvature = .3, yend = .75) +
stamp_polygon(x0 = .2, y0 = .6, radius = .2, alpha = .4) +
stamp_text(x = .2, y = .6, label = "{ggstamp}", size = 5, color = "grey97")

last_plot() +
theme_void()
}
