% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdraft.R
\name{ggdraft}
\alias{ggdraft}
\title{Create a drafting space of fixed width and height}
\usage{
ggdraft(x = 0, y = 0, width = 1, height = width/1.618)
}
\arguments{
\item{x}{numeric x origin of drafting space}

\item{y}{numeric y origin of drafting space}

\item{width}{numeric width of drafting space, defaults to 1}

\item{height}{numeric heigh of drafting space, defaults width/1.618}
}
\value{

}
\description{
Create a drafting space of fixed width and height
}
\examples{
library(ggplot2)
ggdraft() +
stamp_curve(curvature = .13,
            size = .09,
            yend = 20:60/100,
            xend = 1) +
stamp_polygon(x0 = .20, y0 = .40,
              radius = .20, alpha = .4) +
stamp_text(x = .20, y = .40, label = "{ggstamp}",
           size = 6, color = "grey97")

# remove drafting lines
(ggdraft() +
stamp_polygon(x0 = .20, y0 = .40,
              radius = .20, alpha = .4) +
stamp_polygon(x0 = .85, y0 = .23,
              radius = .22,
              n = 5, fill = "seagreen2")
   ) +
theme_void()
}
