% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_honeycomb_y.R
\name{pos_honeycomb_y}
\alias{pos_honeycomb_y}
\title{Title}
\usage{
pos_honeycomb_y(y0 = 0, n = 5, ncol = 3, height = 1.5)
}
\arguments{
\item{height}{}
}
\value{

}
\description{
Title
}
\examples{
pos_honeycomb_y()
ggcanvas() +
 stamp_polygon(x0 = pos_honeycomb_x(),
               y0 = pos_honeycomb_y())
}
