% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_wrap_png_x.R
\name{pos_wrap_x_png}
\alias{pos_wrap_x_png}
\title{Title}
\usage{
pos_wrap_x_png(
  x0 = 0,
  n = 5,
  ncol = 3,
  png = system.file("img", "Rlogo.png", package = "png"),
  image = png::readPNG(png),
  width = 1
)
}
\arguments{
\item{n}{}
}
\value{

}
\description{
Title
}
