% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_arrow.R
\name{stamp_arrow}
\alias{stamp_arrow}
\title{Stamp an arrow/arrows}
\usage{
stamp_arrow(
  x = 0,
  xend = 1,
  y = 0,
  yend = 1,
  size = 1.5,
  alpha = 1,
  color = "black",
  ends = "last",
  units = "npc",
  headlength = 0.03,
  angle = 30,
  type = "open"
)
}
\arguments{
\item{x}{defaults to 0}

\item{xend}{defaults to 1}

\item{y}{defaults to 0}

\item{yend}{defaults to 1}

\item{size}{numeric indicating line width}

\item{alpha}{numeric between 0 and 1 indicating transparency}

\item{ends}{One of "last", "first", or "both", indicating which ends of the line to draw arrow heads.}

\item{units}{A character vector specifying the units for the corresponding numeric values.}

\item{headlength}{A unit specifying the length of the arrow head (from tip to base).}

\item{angle}{The angle of the arrow head in degrees (smaller numbers produce narrower, pointier arrows). Essentially describes the width of the arrow head.}

\item{type}{One of "open" or "closed" indicating whether the arrow head should be a closed triangle.}
}
\value{

}
\description{
This function adds an arrow annotation layer
Contribute or help document https://github.com/EvaMaeRey/ggstamp/blob/master/R/stamp_arrow.R
}
\examples{
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_arrow(color = "darkred",
 xend = 15, yend = 75)

 ggcanvas() +
 stamp_arrow(color = "steelblue") +
 stamp_arrow(x = 1:5, y = 2, xend = 2:6,
 headlength = .04, type = "closed", color = "darkred") +
 stamp_arrow(x = .5, xend = 1.5, ends = "both", angle = 60)
}
