% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_arrow.R
\name{stamp_arrow}
\alias{stamp_arrow}
\title{Stamp some text}
\usage{
stamp_arrow(
  x = 0,
  xend = 1,
  y = 0,
  yend = 1,
  size = 1.5,
  alpha = 1,
  color = "black",
  ends = "last",
  units = "npc",
  headlength = 0.03,
  angle = 30,
  type = "open"
)
}
\arguments{
\item{x}{defaults to 0}

\item{y}{defaults to 0}

\item{size}{}

\item{label}{defaults with a message about used}
}
\value{

}
\description{
This function adds a text annotation layer
}
\examples{
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_arrow(color = "darkred",
 xend = 15, yend = 75)

 ggcanvas() +
 stamp_arrow(color = "steelblue") +
 stamp_arrow(x = 1:5, y = 2, xend = 2:6,
 headlength = .04, type = "closed", color = "darkred") +
 stamp_arrow(x = .5, xend = 1.5, ends = "both", angle = 60)
}
