% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_curve.R
\name{stamp_curve}
\alias{stamp_curve}
\title{Stamp some text}
\usage{
stamp_curve(
  x = 0,
  xend = 1,
  y = 0,
  yend = 1,
  alpha = 1,
  color = "black",
  linetype = "solid",
  size = 1.5,
  curvature = 0.1
)
}
\arguments{
\item{x}{defaults to 0}

\item{xend}{numeric defaults to 1}

\item{y}{defaults to 0}

\item{yend}{numeric defaults to 1}

\item{alpha}{a numeric between 0 and 1 for transparency}

\item{linetype}{"solid", "dashed", "dotted"}

\item{size}{numeric indicating line width, defaults to 1.5}

\item{curvature}{a scalar between -1 and 1, 0 is straight, positive is counter clockwise?}
}
\value{

}
\description{
This function adds a text annotation layer
}
\examples{
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_curve(color = "darkred",
 xend = 15, yend = 75)

 ggcanvas() +
 stamp_curve(color = "green") +
 stamp_curve(x = 1:5, yend = 2, xend = 2:6,
 size = 8, alpha = .5) +
 stamp_point(x = 1:5, y = 0)
}
