% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_ellipse.R
\name{stamp_ellipse}
\alias{stamp_ellipse}
\title{Stamp some text}
\usage{
stamp_ellipse(
  x0 = 0,
  y0 = 0,
  a = 1,
  b = 0.5,
  angle = 0,
  size = 1.5,
  alpha = 1,
  linetype = "solid",
  fill = "grey85",
  color = "black",
  x0y0 = NULL
)
}
\arguments{
\item{x0}{defaults to 0}

\item{y0}{defaults to 0}

\item{size}{}

\item{label}{defaults with a message about used}
}
\value{

}
\description{
This function adds a text annotation layer
}
\examples{
# an ellipse
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 stamp_ellipse() +
 stamp_ellipse(color = "plum4", x0 = 2, y0 = 3, angle = pi/6)

# several
ggcanvas() +
 stamp_ellipse(color = "green") +
 stamp_ellipse(x0 = 1:5, y0 = 2, size = 2)
}
