% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_grid.R
\name{stamp_grid}
\alias{stamp_grid}
\title{Title}
\usage{
stamp_grid(color = "turquoise4", fill = "linen", alpha = 0.5)
}
\value{

}
\description{
Title
}
\examples{
library(ggplot2)
ggplot(cars) +
  aes(x = speed) +
  aes(y = dist) +
  geom_point() +
  facet_wrap(~speed > 15) ->
  p

cowplot::ggdraw(p) +
  stamp_grid() + # you can comment this guide out later
  stamp_text_ljust(label = "Over on this side we look at the slow cars." \%>\%
  stringr::str_wrap(15), x = .15, y = .8, size = 5) +
  stamp_segment(x = .35, xend = .35, y = .2, yend = .7,
                size = .5, linetype = "dashed") +
  stamp_text_ljust(label = "Over here, fast cars!" \%>\%
  stringr::str_wrap(15), x = .6, y = .2, size = 5) +
  stamp_curve(x = .65, y = .27, xend = .79, yend = .5, curvature = -.2)

}
