% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_heart.R
\name{stamp_heart}
\alias{stamp_heart}
\title{Stamp a heart of n sides}
\usage{
stamp_heart(
  x0 = 0,
  y0 = 0,
  n = 200,
  radius = 1,
  size = 1.5,
  alpha = 1,
  rotation = 0,
  linetype = "solid",
  fill = "grey35",
  color = "black",
  x0y0 = NULL
)
}
\arguments{
\item{x0}{defaults to 0}

\item{y0}{defaults to 0}

\item{color}{}
}
\value{

}
\description{
This function adds a text annotation layer.
Contribute or help document https://github.com/EvaMaeRey/ggstamp/blob/master/R/stamp_polygon.R
}
\examples{
# on plot space
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_heart(color = "darkred",
             fill = "pink")

 ggcanvas() +
  stamp_heart()

# on a canvas
 ggcanvas() +
  stamp_heart(n = 18, x0 = c(0, 2.5),
  fill = "magenta", color = "purple") +
  stamp_label(x = pos_spoke_x(x0 = 0, n = 3),
              y = pos_spoke_y(y0 = 0, n = 3),
              label = c("hi", "hello", "bye")) +
  stamp_point(color = "turquoise")

# lots of polygons
 ggcanvas() +
 stamp_heart(color = "turquoise4", radius = 3, fill = "turquoise") +
 stamp_heart(fill = "slateblue2") +
 stamp_heart(linetype = "dashed", alpha = .5, fill = "plum2") +
 stamp_heart(x0 = 1, y0 = 1.5, n = 3, rotation = .25, size = 0,
  fill = "goldenrod2") +
 stamp_heart(n = 50, radius = .2, size = 0) +
 stamp_spoke(angle = pi * 1:7/15, radius = 2)

# some more
 ggcanvas() +
  stamp_heart(x0y0 = pos_wrap(n = 5, ncol = 3, height = 1.5),
                radius = .5) +
  stamp_text(xy = pos_wrap(n = 5, ncol = 3,
           height = 1.5), label = LETTERS[1:5])

}
