% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_hline.R
\name{stamp_hline}
\alias{stamp_hline}
\title{Horizontal line, wraps geom_hline w defaut at y = 0}
\usage{
stamp_hline(
  yintercept = 0,
  alpha = 1,
  color = "black",
  linetype = "solid",
  size = 1.5
)
}
\arguments{
\item{size}{}
}
\value{

}
\description{
Horizontal line, wraps geom_hline w defaut at y = 0
}
\examples{
library(ggplot2)
ggplot(cars) +
aes(speed, dist) +
stamp_hline()

# on a canvas
ggcanvas() +
stamp_hline() +
stamp_hline(1:6,
           linetype = rep(c("dotted", "dashed"), 3))
}
