% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_label.R
\name{stamp_label}
\alias{stamp_label}
\title{Stamp some text}
\usage{
stamp_label(
  x = 0,
  y = 0,
  label = "use label\\nargument to change\\nthis label",
  size = 8,
  alpha = 1,
  angle = 0,
  color = "black",
  fill = "white",
  lineheight = 0.8,
  family = "",
  vjust = 0.5,
  label.padding = ggplot2::unit(0.5, units = "lines"),
  ...
)
}
\arguments{
\item{x}{defaults to 0}

\item{y}{defaults to 0}

\item{label}{defaults with a message about use}

\item{size}{}
}
\value{

}
\description{
This function adds a text annotation layer
}
\examples{
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_label() +
 stamp_label(label = "hi", y = c(100,75,50), x = 20)
}
