% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_path.R
\name{stamp_path}
\alias{stamp_path}
\title{Stamp a path, connected points}
\usage{
stamp_path(
  x = 0:3,
  y = c(0, 1, 0, 1),
  alpha = 1,
  color = "black",
  linetype = "solid",
  size = 1.5,
  lineend = "butt",
  linejoin = "round",
  arrow = ggplot2::arrow(length = 0),
  arrow.fill = NULL,
  xy = NULL
)
}
\arguments{
\item{x}{numeric x position, defaults to 0}

\item{y}{numeric y position, defaults to 0}

\item{alpha}{numeric between 0 and 1 indicating transparency, defaults to 1 fully opaque, zero is fully transparent}

\item{color}{a character string indicating color, defaults to "black", for other named colors check out https://evamaerey.github.io/ggplot2_grammar_guide/named_colors}

\item{linetype}{a string indicating linetype, "solid", "dashed", "dotted", defaults to "solid"}

\item{size}{numeric value indicating line width, defaults to 1.5}

\item{lineend}{Line end style (round, butt, square)}

\item{arrow}{specification for arrow heads, as created by arrow().}
}
\value{

}
\description{
This function adds a text annotation layer
}
\examples{
# segments on canvas
 ggcanvas() +
 stamp_path(color = "green") +
 stamp_path(xy = pos_honeycomb(ncol = 1, n = 12),
            alpha = .5, lineend = "round")


}
