% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_png.R
\name{stamp_png}
\alias{stamp_png}
\title{stamp_png}
\usage{
stamp_png(
  x0 = 0,
  y0 = 0,
  png = system.file("img", "Rlogo.png", package = "png"),
  image = png::readPNG(png),
  width = 1,
  height = width * dim(image)[2]/dim(image)[1],
  show.legend = F,
  x0y0 = NULL
)
}
\arguments{
\item{ymax}{}
}
\value{

}
\description{
stamp_png
}
\examples{
library(ggplot)
ggplot(cars, aes(dist, speed)) +
  geom_point() +
  stamp_png(x0 = 5:11*10, y0 = 10, height = 1, width = 10) +
  scale_fill_identity()

ggcanvas() +
  stamp_png() +
  stamp_png(x0 = 2:3, y0 = 6, width = .95) +
  scale_fill_identity()

ggcanvas() +
  stamp_png(x0 = 0:8 \%\% 5 + 1,
        y = -(0:8 \%/\% 5)/.75 , width = .90) +
  coord_equal() +
  stamp_png(x = 0:5 \%\% 5 + 1,
      y = -(0:5 \%/\% 5)/.75  -3, width = .90) +
  stamp_png(x = 0:2 \%\% 5 + 1,
     y = -(0:2 \%/\% 5)/.75  -6, width = .90) +
  scale_fill_identity()

ggcanvas() +
  stamp_png(x0 = wrap_x(),
            y0 = wrap_y_png())
}
