% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_png.R
\name{stamp_png}
\alias{stamp_png}
\title{stamp_png}
\usage{
stamp_png(
  x0 = 0,
  y0 = 0,
  png = system.file("img", "Rlogo.png", package = "png"),
  image = png::readPNG(png),
  width = 1,
  height = width * dim(image)[1]/dim(image)[2],
  show.legend = F,
  alpha = 1,
  x0y0 = NULL
)
}
\arguments{
\item{ymax}{}
}
\value{

}
\description{
stamp_png
}
\examples{
# stamp in plotspace
library(ggplot2)
ggplot(cars, aes(dist, speed)) +
  geom_point() +
  stamp_png(x0 = 5:11*10, y0 = 10,
            height = 1, width = 10) +
  scale_fill_identity()

# two stamps
ggcanvas() +
  stamp_png() +
  stamp_png(x0 = 2:3, y0 = 6,
            width = .95) +
  scale_fill_identity()

# png spokes
ggcanvas() +
  stamp_png()

# tiling pngs needs work
ggcanvas() +
  stamp_png(x0y0 = wrap_png(),
  alpha = .75)
}
