% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_point.R
\name{stamp_point}
\alias{stamp_point}
\title{Stamp a point/points}
\usage{
stamp_point(
  x = 0,
  y = 0,
  size = 8,
  alpha = 1,
  shape = "circle",
  color = "black",
  stroke = 1,
  fill = "grey35",
  xy = NULL
)
}
\arguments{
\item{x}{a scalar or vector numeric}

\item{y}{a scalar or vector numeric}

\item{size}{a scalar or vector numeric}

\item{alpha}{a numeric between 0 and 1 for transparency}

\item{shape}{a numeric or character string, "circle" is default but also try "star", "bullet", "diamond", "triangle"}

\item{color}{a color as a character string or hex, named colors: https://evamaerey.github.io/ggplot2_grammar_guide/named_colors}

\item{stroke}{a numeric - size of shape outline}

\item{fill}{a fill color as a character string or hex, named colors: https://evamaerey.github.io/ggplot2_grammar_guide/named_colors}
}
\value{

}
\description{
This function adds a point annotation layer
}
\examples{
# stamp in plot space
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_point() +
 stamp_point(color = "red", x = 10, y = 100)

# more points on canvas
ggcanvas() +
 stamp_point(color = "darkgreen") +
 stamp_point(x = 1:5, y = 2,
             size = 8, alpha = .5) +
 stamp_point(xy = pos_wrap(y0 = -2,
             n = 13, ncol = 4),
             color = "plum") +
 stamp_point(xy = pos_wrap(x0 = 1,
             n = 7, ncol = 4, width = .5))
}
