% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_polygon_holes.R
\name{stamp_polygon_holes}
\alias{stamp_polygon_holes}
\title{Stamp a polygon of n sides}
\usage{
stamp_polygon_holes(
  x0 = 0,
  y0 = 0,
  n = 6,
  radius = 1,
  radius_outer = Inf,
  size = 0,
  alpha = 1,
  rotation = -0.5,
  linetype = "solid",
  fill = "white",
  color = "black",
  x0y0 = NULL
)
}
\arguments{
\item{x0}{defaults to 0}

\item{y0}{defaults to 0}

\item{color}{}
}
\value{

}
\description{
This function adds a text annotation layer.
Contribute or help document https://github.com/EvaMaeRey/ggstamp/blob/master/R/stamp_polygon.R
}
\examples{
# some more
 ggcanvas() +
  stamp_png(width = 3) +
  stamp_polygon_holes(x0y0 = pos_honeycomb(n = 5),
  radius = .95,
                        radius_outer = Inf) +
  stamp_polygon(alpha = .2, radius = .95,
                color = "grey30")
}
