% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_segment.R
\name{stamp_segment}
\alias{stamp_segment}
\title{Stamp some text}
\usage{
stamp_segment(
  x = 0,
  xend = 1,
  y = 0,
  yend = 1,
  size = 1.5,
  alpha = 1,
  color = "black"
)
}
\arguments{
\item{x}{defaults to 0}

\item{y}{defaults to 0}

\item{size}{}

\item{label}{defaults with a message about used}
}
\value{

}
\description{
This function adds a text annotation layer
}
\examples{
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_segment(color = "darkred",
 xend = 15, yend = 75)

 ggcanvas() +
 stamp_segment(color = "green") +
 stamp_segment(x = 1:5, y = 2, xend = 2:6,
 size = 8, alpha = .5)
}
