% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_text.R
\name{stamp_text}
\alias{stamp_text}
\title{Stamp some text}
\usage{
stamp_text(
  x = 0,
  y = 0,
  label = "use label\\nargument to change\\nthis text",
  size = 8,
  alpha = 1,
  angle = 0,
  color = "black",
  lineheight = 0.8,
  family = "",
  vjust = 0.5,
  ...
)
}
\arguments{
\item{x}{defaults to 0}

\item{y}{defaults to 0}

\item{label}{defaults with a message about used}

\item{size}{}
}
\value{

}
\description{
This function adds a text annotation layer
}
\examples{
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_text() +
 stamp_text(label = "hi", x = 15:18, y = 75)
}
