% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_tile.R
\name{stamp_tile}
\alias{stamp_tile}
\title{Stamp some text}
\usage{
stamp_tile(
  x = 0,
  y = 0,
  size = 0.5,
  alpha = 1,
  color = "black",
  fill = "grey85",
  height = 1,
  linetype = "solid",
  width = 1,
  xy = NULL
)
}
\arguments{
\item{x}{defaults to 0}

\item{y}{defaults to 0}

\item{size}{line width default is .5}

\item{alpha}{0 to 1 indicating transparency}

\item{color}{named or hexcolor for outline of tile}

\item{fill}{background color of label, consider named colors https://evamaerey.github.io/ggplot2_grammar_guide/named_colors}

\item{height}{full height of tile centered at y}

\item{linetype}{"solid" "dashed" "dotted" linetype border}

\item{width}{full width of tile centered at x}

\item{xy}{a data frame where the first two columns contain x and y positional values}
}
\value{

}
\description{
This function adds a text annotation layer
}
\examples{
# tile
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_tile(color = "red", x = 10,
            y = 90:100) +
 stamp_tile(height = 10,
            xy = spoke(x0 = 20, n = 15,
            radius = 6))

# on a canvas
ggcanvas() +
 stamp_tile(fill = "green") +
 stamp_tile(x = 1:5, y = 2, alpha = .8,
            fill = "steelblue",
            width = .95,
            size = 0) +
 coord_equal()

# using a positioner
ggcanvas() +
 stamp_tile(xy = spoke())
}
