% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_wash.R
\name{stamp_wash}
\alias{stamp_wash}
\title{Stamp some text}
\usage{
stamp_wash(
  xmin = -Inf,
  ymin = -Inf,
  xmax = Inf,
  ymax = Inf,
  x = 0,
  y = 0,
  alpha = 0.5,
  color = alpha("linen", 0),
  fill = "linen",
  linetype = "solid",
  size = 0.5,
  xxyy = NULL
)
}
\value{

}
\description{
This function adds a text annotation layer
}
\examples{
# wash
library(ggplot2)
ggplot(cars) +
 aes(speed, dist) +
 geom_point() +
 stamp_wash() +
 stamp_wash(fill = "red")

# over backdrop
ggbackdrop() +
 stamp_wash(fill = "green",
            alpha = .2)
}
