% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_void_fill.R
\name{theme_void_fill}
\alias{theme_void_fill}
\title{Title}
\usage{
theme_void_fill(
  fill = "linen",
  color = fill,
  margin.top = 10,
  margin.bottom = margin.top,
  margin.left = margin.top,
  margin.right = margin.top,
  ...
)
}
\arguments{
\item{fill}{string indicating background color, defaults to "linen"}

\item{color}{string indicating seams color, defaults to fill color}
}
\value{

}
\description{
Title
}
\examples{
ggcanvas() +
  stamp_png() +
  theme_void_fill()
}
