% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_png.R
\name{wrap_png}
\alias{wrap_png}
\title{Title}
\usage{
wrap_png(
  x0 = 0,
  y0 = 0,
  n = 5,
  ncol = 3,
  png = system.file("img", "Rlogo.png", package = "png"),
  image = png::readPNG(png),
  width = 1,
  height = width * dim(image)[2]/dim(image)[1]
)
}
\arguments{
\item{height}{}
}
\value{

}
\description{
Title
}
\examples{
wrap_png()
ggcanvas() +
stamp_png(x0y0 = wrap_png(width = 1.2)) +
stamp_png(x0y0 = wrap_png(width = 1.2, y0 = -4))
}
